requirements.txt
requests>=2.31.0
python-telegram-bot>=20.0
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=0.19.0
websockets>=10.0
ccxt>=4.1.13
aiosqlite>=0.17.0
ujson>=5.0
aiodns>=3.0.0
charset-normalizer>=2.1.0
orjson>=3.6.0
uvloop>=0.16.0; sys_platform != "win32"
aiofiles>=0.8.0
tenacity>=8.0.0
prometheus_client>=0.12.0
SQLAlchemy>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
alembic>=1.12.0 